+++ NEWS.md
@@ -44,6 +44,9 @@
 
 * Add support for `as.Date()` for character columns (#190).
 
+* Error messages due to untranslated functions now suggest opening an issue to
+  ask for their translation (#197).
+
 ## Bug fixes
 
 * Fix behavior of `mutate()` and `summarize()` when they don't contain any